Refactor BaseAsyncHttpConnector to use TAP
Also moved BaseAsyncHttpConnector class into its own source file.
