+++ Libraries/dotNetRDF/Storage/BaseHttpConnector.cs
@@ -25,29 +25,16 @@
 */
 
 using System;
-using System.Collections.Generic;
-using System.IO;
 using System.Net;
 using System.Text;
-using System.Threading;
 using VDS.RDF.Configuration;
 using VDS.RDF.Parsing;
-using VDS.RDF.Parsing.Handlers;
-using VDS.RDF.Storage.Management;
 
 namespace VDS.RDF.Storage
 {
     /// <summary>
     /// Abstract Base Class for HTTP based Storage API implementations.
     /// </summary>
-    /// <remarks>
-    /// <para>
-    /// As .NET Core does not currently provide an implementaton of the System.Net.WebProxy class, this implementation
-    /// is restricted in functionality. Many of the methods it exposes will raise a PlatformNotSupported exception,
-    /// but are currently required to enable dependent classes to compile. As such it is strongly recommended that
-    /// developers targeting .NET Core should NOT plan on making any use of the proxy capabilities of this class.
-    /// </para>
-    /// </remarks>
     public abstract class BaseHttpConnector
     {
         /// <summary>
@@ -63,15 +50,13 @@ namespace VDS.RDF.Storage
         /// </summary>
         private bool _hasCredentials;
 
-        private IWebProxy _proxy;
-
         /// <summary>
         /// Sets a Proxy Server to be used.
         /// </summary>
         /// <param name="address">Proxy Address.</param>
-        public void SetProxy(String address)
+        public void SetProxy(string address)
         {
-            _proxy = new WebProxy(address);
+            Proxy = new WebProxy(address);
         }
 
         /// <summary>
@@ -80,30 +65,20 @@ namespace VDS.RDF.Storage
         /// <param name="address">Proxy Address.</param>
         public void SetProxy(Uri address)
         {
-            _proxy = new WebProxy(address);
+            Proxy = new WebProxy(address);
         }
 
         /// <summary>
         /// Gets/Sets a Proxy Server to be used.
         /// </summary>
-        public IWebProxy Proxy
-        {
-            get
-            {
-                return _proxy;
-            }
-            set
-            {
-                _proxy = value;
-            }
-        }
+        public IWebProxy Proxy { get; set; }
 
         /// <summary>
         /// Clears any in-use credentials so subsequent requests will not use a proxy server.
         /// </summary>
         public void ClearProxy()
         {
-            _proxy = null;
+            Proxy = null;
         }
 
         /// <summary>
@@ -111,11 +86,11 @@ namespace VDS.RDF.Storage
         /// </summary>
         /// <param name="username">Username.</param>
         /// <param name="password">Password.</param>
-        public void SetProxyCredentials(String username, String password)
+        public void SetProxyCredentials(string username, string password)
         {
-            if (_proxy != null)
+            if (Proxy != null)
             {
-                _proxy.Credentials = new NetworkCredential(username, password);
+                Proxy.Credentials = new NetworkCredential(username, password);
             }
             else
             {
@@ -129,11 +104,11 @@ namespace VDS.RDF.Storage
         /// <param name="username">Username.</param>
         /// <param name="password">Password.</param>
         /// <param name="domain">Domain.</param>
-        public void SetProxyCredentials(String username, String password, String domain)
+        public void SetProxyCredentials(string username, string password, string domain)
         {
-            if (_proxy != null)
+            if (Proxy != null)
             {
-                _proxy.Credentials = new NetworkCredential(username, password, domain);
+                Proxy.Credentials = new NetworkCredential(username, password, domain);
             }
             else
             {
@@ -146,22 +121,12 @@ namespace VDS.RDF.Storage
         /// </summary>
         public ICredentials ProxyCredentials
         {
-            get
-            {
-                if (_proxy != null)
-                {
-                    return _proxy.Credentials;
-                }
-                else
-                {
-                    return null;
-                }
-            }
+            get => Proxy?.Credentials;
             set
             {
-                if (_proxy != null)
+                if (Proxy != null)
                 {
-                    _proxy.Credentials = value;
+                    Proxy.Credentials = value;
                 }
                 else
                 {
@@ -175,9 +140,9 @@ namespace VDS.RDF.Storage
         /// </summary>
         public void ClearProxyCredentials()
         {
-            if (_proxy != null)
+            if (Proxy != null)
             {
-                _proxy.Credentials = null;
+                Proxy.Credentials = null;
             }
         }
 
@@ -218,9 +183,9 @@ namespace VDS.RDF.Storage
         protected HttpWebRequest ApplyRequestOptions(HttpWebRequest request)
         {
             if (Timeout > 0) request.Timeout = Timeout;
-            if (_proxy != null)
+            if (Proxy != null)
             {
-                request.Proxy = _proxy;
+                request.Proxy = Proxy;
             }
 
             // Add Credentials if needed
@@ -269,7 +234,7 @@ namespace VDS.RDF.Storage
             }
 
             // Proxy configuration
-            if (_proxy == null) return;
+            if (Proxy == null) return;
             INode proxy = context.NextSubject;
             INode usesProxy = context.Graph.CreateUriNode(UriFactory.Create(ConfigurationLoader.PropertyProxy));
             INode rdfType = context.Graph.CreateUriNode(UriFactory.Create(RdfSpecsHelper.RdfType));
@@ -280,10 +245,10 @@ namespace VDS.RDF.Storage
 
             context.Graph.Assert(new Triple(objNode, usesProxy, proxy));
             context.Graph.Assert(new Triple(proxy, rdfType, proxyType));
-            context.Graph.Assert(new Triple(proxy, server, context.Graph.CreateLiteralNode((_proxy as WebProxy).Address.AbsoluteUri)));
+            context.Graph.Assert(new Triple(proxy, server, context.Graph.CreateLiteralNode((Proxy as WebProxy).Address.AbsoluteUri)));
 
-            if (!(_proxy.Credentials is NetworkCredential)) return;
-            NetworkCredential cred = (NetworkCredential)_proxy.Credentials;
+            if (!(Proxy.Credentials is NetworkCredential)) return;
+            NetworkCredential cred = (NetworkCredential)Proxy.Credentials;
             context.Graph.Assert(new Triple(proxy, user, context.Graph.CreateLiteralNode(cred.UserName)));
             context.Graph.Assert(new Triple(proxy, pwd, context.Graph.CreateLiteralNode(cred.Password)));
         }
@@ -292,486 +257,7 @@ namespace VDS.RDF.Storage
         {
             Username = username;
             Password = password;
-            _hasCredentials = (!String.IsNullOrEmpty(username) && !String.IsNullOrEmpty(password));
-        }
-    }
-
-    /// <summary>
-    /// Abstract Base Class for HTTP Based <see cref="IAsyncStorageProvider">IAsyncStorageProvider</see> implementations.
-    /// </summary>
-    /// <remarks>
-    /// <para>
-    /// It is expected that most classes extending from this will also then implement <see cref="IStorageProvider"/> separately for their synchronous communication, this class purely provides partial helper implementations for the asynchronous communication.
-    /// </para>
-    /// </remarks>
-    public abstract class BaseAsyncHttpConnector
-        : BaseHttpConnector, IAsyncStorageProvider
-    {
-        private readonly DoRequestSequenceDelgate _d;
-
-        /// <summary>
-        /// Creates a new Base Async HTTP Connector.
-        /// </summary>
-        protected BaseAsyncHttpConnector()
-        {
-            _d = new DoRequestSequenceDelgate(DoRequestSequence);
-        }
-
-        /// <summary>
-        /// Gets the parent server (if any).
-        /// </summary>
-        public virtual IStorageServer ParentServer
-        {
-            get
-            {
-                return null;
-            }
-        }
-
-        /// <summary>
-        /// Gets the parent server (if any).
-        /// </summary>
-        public virtual IAsyncStorageServer AsyncParentServer
-        {
-            get
-            {
-                return null;
-            }
-        }
-
-        /// <summary>
-        /// Loads a Graph from the Store asynchronously.
-        /// </summary>
-        /// <param name="g">Graph to load into.</param>
-        /// <param name="graphUri">URI of the Graph to load.</param>
-        /// <param name="callback">Callback.</param>
-        /// <param name="state">State to pass to the callback.</param>
-        public virtual void LoadGraph(IGraph g, Uri graphUri, AsyncStorageCallback callback, Object state)
-        {
-            LoadGraph(g, graphUri.ToSafeString(), callback, state);
-        }
-
-        /// <summary>
-        /// Loads a Graph from the Store asynchronously.
-        /// </summary>
-        /// <param name="g">Graph to load into.</param>
-        /// <param name="graphUri">URI of the Graph to load.</param>
-        /// <param name="callback">Callback.</param>
-        /// <param name="state">State to pass to the callback.</param>
-        public virtual void LoadGraph(IGraph g, String graphUri, AsyncStorageCallback callback, Object state)
-        {
-            LoadGraph(new GraphHandler(g), graphUri, (sender, args, st) => callback(sender, new AsyncStorageCallbackArgs(AsyncStorageOperation.LoadGraph, g, args.Error), st), state);
-        }
-
-        /// <summary>
-        /// Loads a Graph from the Store asynchronously.
-        /// </summary>
-        /// <param name="handler">Handler to load with.</param>
-        /// <param name="graphUri">URI of the Graph to load.</param>
-        /// <param name="callback">Callback.</param>
-        /// <param name="state">State to pass to the callback.</param>
-        public virtual void LoadGraph(IRdfHandler handler, Uri graphUri, AsyncStorageCallback callback, Object state)
-        {
-            LoadGraph(handler, graphUri.ToSafeString(), callback, state);
-        }
-
-        /// <summary>
-        /// Loads a Graph from the Store asynchronously.
-        /// </summary>
-        /// <param name="handler">Handler to load with.</param>
-        /// <param name="graphUri">URI of the Graph to load.</param>
-        /// <param name="callback">Callback.</param>
-        /// <param name="state">State to pass to the callback.</param>
-        public abstract void LoadGraph(IRdfHandler handler, String graphUri, AsyncStorageCallback callback, Object state);
-
-        /// <summary>
-        /// Helper method for doing async load operations, callers just need to provide an appropriately prepared HTTP request.
-        /// </summary>
-        /// <param name="request">HTTP Request.</param>
-        /// <param name="handler">Handler to load with.</param>
-        /// <param name="callback">Callback.</param>
-        /// <param name="state">State to pass to the callback.</param>
-        protected internal void LoadGraphAsync(HttpWebRequest request, IRdfHandler handler, AsyncStorageCallback callback, Object state)
-        {
-            Tools.HttpDebugRequest(request);
-            request.BeginGetResponse(r =>
-            {
-                try
-                {
-                    HttpWebResponse response = (HttpWebResponse)request.EndGetResponse(r);
-                    Tools.HttpDebugResponse(response);
-                    // Parse the retrieved RDF
-                    IRdfReader parser = MimeTypesHelper.GetParser(response.ContentType);
-                    parser.Load(handler, new StreamReader(response.GetResponseStream()));
-
-                    // If we get here then it was OK
-                    response.Close();
-
-                    callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.LoadWithHandler, handler), state);
-                }
-                catch (WebException webEx)
-                {
-                    if (webEx.Response != null)
-                    {
-                        Tools.HttpDebugResponse((HttpWebResponse)webEx.Response);
-                        if (((HttpWebResponse)webEx.Response).StatusCode == HttpStatusCode.NotFound)
-                        {
-                            callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.LoadWithHandler, handler), state);
-                            return;
-                        }
-                    }
-                    callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.LoadWithHandler, handler, new RdfStorageException("A HTTP Error occurred while trying to load a Graph from the Store", webEx)), state);
-                }
-                catch (Exception ex)
-                {
-                    callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.LoadWithHandler, handler, StorageHelper.HandleError(ex, "loading a Graph asynchronously from")), state);
-                }
-            }, state);
-        }
-
-        /// <summary>
-        /// Saves a Graph to the Store asynchronously.
-        /// </summary>
-        /// <param name="g">Graph to save.</param>
-        /// <param name="callback">Callback.</param>
-        /// <param name="state">State to pass to the callback.</param>
-        public abstract void SaveGraph(IGraph g, AsyncStorageCallback callback, Object state);
-
-        /// <summary>
-        /// Helper method for doing async save operations, callers just need to provide an appropriately perpared HTTP requests and a RDF writer which will be used to write the data to the request body.
-        /// </summary>
-        /// <param name="request">HTTP request.</param>
-        /// <param name="writer">RDF Writer.</param>
-        /// <param name="g">Graph to save.</param>
-        /// <param name="callback">Callback.</param>
-        /// <param name="state">State to pass to the callback.</param>
-        protected internal void SaveGraphAsync(HttpWebRequest request, IRdfWriter writer, IGraph g, AsyncStorageCallback callback, Object state)
-        {
-            request.BeginGetRequestStream(r =>
-            {
-                try
-                {
-                    Stream reqStream = request.EndGetRequestStream(r);
-                    writer.Save(g, new StreamWriter(reqStream));
-
-                    Tools.HttpDebugRequest(request);
-                    
-                    request.BeginGetResponse(r2 =>
-                    {
-                        try
-                        {
-                            HttpWebResponse response = (HttpWebResponse)request.EndGetResponse(r2);
-                            Tools.HttpDebugResponse(response);
-                            // If we get here then it was OK
-                            response.Close();
-                            callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.SaveGraph, g), state);
-                        }
-                        catch (WebException webEx)
-                        {
-                            callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.SaveGraph, g, StorageHelper.HandleHttpError(webEx, "saving a Graph asynchronously to")), state);
-                        }
-                        catch (Exception ex)
-                        {
-                            callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.SaveGraph, g, StorageHelper.HandleError(ex, "saving a Graph asynchronously to")), state);
-                        }
-                    }, state);
-                }
-                catch (WebException webEx)
-                {
-                    callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.SaveGraph, g, StorageHelper.HandleHttpError(webEx, "saving a Graph asynchronously to")), state);
-                }
-                catch (Exception ex)
-                {
-                    callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.SaveGraph, g, StorageHelper.HandleError(ex, "saving a Graph asynchronously to")), state);
-                }
-            }, state);
-        }
-
-        /// <summary>
-        /// Updates a Graph in the Store asychronously.
-        /// </summary>
-        /// <param name="graphUri">URI of the Graph to update.</param>
-        /// <param name="additions">Triples to be added.</param>
-        /// <param name="removals">Triples to be removed.</param>
-        /// <param name="callback">Callback.</param>
-        /// <param name="state">State to pass to the callback.</param>
-        public virtual void UpdateGraph(Uri graphUri, IEnumerable<Triple> additions, IEnumerable<Triple> removals, AsyncStorageCallback callback, Object state)
-        {
-            UpdateGraph(graphUri.ToSafeString(), additions, removals, callback, state);
-        }
-
-        /// <summary>
-        /// Updates a Graph in the Store asychronously.
-        /// </summary>
-        /// <param name="graphUri">URI of the Graph to update.</param>
-        /// <param name="additions">Triples to be added.</param>
-        /// <param name="removals">Triples to be removed.</param>
-        /// <param name="callback">Callback.</param>
-        /// <param name="state">State to pass to the callback.</param>
-        public abstract void UpdateGraph(String graphUri, IEnumerable<Triple> additions, IEnumerable<Triple> removals, AsyncStorageCallback callback, Object state);
-
-        /// <summary>
-        /// Helper method for doing async update operations, callers just need to provide an appropriately prepared HTTP request and a RDF writer which will be used to write the data to the request body.
-        /// </summary>
-        /// <param name="request">HTTP Request.</param>
-        /// <param name="writer">RDF writer.</param>
-        /// <param name="graphUri">URI of the Graph to update.</param>
-        /// <param name="ts">Triples.</param>
-        /// <param name="callback">Callback.</param>
-        /// <param name="state">State to pass to the callback.</param>
-        protected internal void UpdateGraphAsync(HttpWebRequest request, IRdfWriter writer, Uri graphUri, IEnumerable<Triple> ts, AsyncStorageCallback callback, Object state)
-        {
-            Graph g = new Graph();
-            g.Assert(ts);
-
-            request.BeginGetRequestStream(r =>
-            {
-                try
-                {
-                    Stream reqStream = request.EndGetRequestStream(r);
-                    writer.Save(g, new StreamWriter(reqStream));
-
-                    Tools.HttpDebugRequest(request);
-
-                    request.BeginGetResponse(r2 =>
-                    {
-                        try
-                        {
-                            HttpWebResponse response = (HttpWebResponse)request.EndGetResponse(r2);
-                            Tools.HttpDebugResponse(response);
-                            // If we get here then it was OK
-                            response.Close();
-                            callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.UpdateGraph, graphUri), state);
-                        }
-                        catch (WebException webEx)
-                        {
-                            callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.UpdateGraph, graphUri, StorageHelper.HandleHttpError(webEx, "updating a Graph asynchronously in")), state);
-                        }
-                        catch (Exception ex)
-                        {
-                            callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.UpdateGraph, graphUri, StorageHelper.HandleError(ex, "updating a Graph asynchronously in")), state);
-                        }
-                    }, state);
-                }
-                catch (WebException webEx)
-                {
-                    callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.UpdateGraph, graphUri, StorageHelper.HandleHttpError(webEx, "updating a Graph asynchronously in")), state);
-                }
-                catch (Exception ex)
-                {
-                    callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.UpdateGraph, graphUri, StorageHelper.HandleError(ex, "updating a Graph asynchronously in")), state);
-                }
-            }, state);
-        }
-
-        /// <summary>
-        /// Deletes a Graph from the Store.
-        /// </summary>
-        /// <param name="graphUri">URI of the Graph to delete.</param>
-        /// <param name="callback">Callback.</param>
-        /// <param name="state">State to pass to the callback.</param>
-        public virtual void DeleteGraph(Uri graphUri, AsyncStorageCallback callback, Object state)
-        {
-            DeleteGraph(graphUri.ToSafeString(), callback, state);
-        }
-
-        /// <summary>
-        /// Deletes a Graph from the Store.
-        /// </summary>
-        /// <param name="graphUri">URI of the Graph to delete.</param>
-        /// <param name="callback">Callback.</param>
-        /// <param name="state">State to pass to the callback.</param>
-        public abstract void DeleteGraph(String graphUri, AsyncStorageCallback callback, Object state);
-
-        /// <summary>
-        /// Helper method for doing async delete operations, callers just need to provide an appropriately prepared HTTP request.
-        /// </summary>
-        /// <param name="request">HTTP request.</param>
-        /// <param name="allow404">Whether a 404 response counts as success.</param>
-        /// <param name="graphUri">URI of the Graph to delete.</param>
-        /// <param name="callback">Callback.</param>
-        /// <param name="state">State to pass to the callback.</param>
-        protected internal void DeleteGraphAsync(HttpWebRequest request, bool allow404, String graphUri, AsyncStorageCallback callback, Object state)
-        {
-            Tools.HttpDebugRequest(request);
-            request.BeginGetResponse(r =>
-                {
-                    try
-                    {
-                        HttpWebResponse response = (HttpWebResponse)request.EndGetResponse(r);
-
-                        // Assume if returns to here we deleted the Graph OK
-                        response.Close();
-                        callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.DeleteGraph, graphUri.ToSafeUri()), state);
-                    }
-                    catch (WebException webEx)
-                    {
-                        if (webEx.Response != null) Tools.HttpDebugResponse((HttpWebResponse)webEx.Response);
-                        // Don't throw the error if we get a 404 - this means we couldn't do a delete as the graph didn't exist to start with
-                        if (webEx.Response == null || (webEx.Response != null && (!allow404 || ((HttpWebResponse)webEx.Response).StatusCode != HttpStatusCode.NotFound)))
-                        {
-                            callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.DeleteGraph, graphUri.ToSafeUri(), new RdfStorageException("A HTTP Error occurred while trying to delete a Graph from the Store asynchronously", webEx)), state);
-                        }
-                        else
-                        {
-                            // Consider a 404 as a success in some cases
-                            callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.DeleteGraph, graphUri.ToSafeUri()), state);
-                        }
-                    }
-                    catch (Exception ex)
-                    {
-                        callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.DeleteGraph, graphUri.ToSafeUri(), StorageHelper.HandleError(ex, "deleting a Graph asynchronously from")), state);
-                    }
-                }, state);
-        }
-
-        /// <summary>
-        /// Lists the Graphs in the Store asynchronously.
-        /// </summary>
-        /// <param name="callback">Callback.</param>
-        /// <param name="state">State to pass to the callback.</param>
-        public virtual void ListGraphs(AsyncStorageCallback callback, object state)
-        {
-            if (this is IAsyncQueryableStorage)
-            {
-                // Use ListUrisHandler and make an async query to list the graphs, when that returns we invoke the correct callback
-                ListUrisHandler handler = new ListUrisHandler("g");
-                ((IAsyncQueryableStorage)this).Query(null, handler, "SELECT DISTINCT ?g WHERE { GRAPH ?g { ?s ?p ?o } }", (sender, args, st) =>
-                {
-                    if (args.WasSuccessful)
-                    {
-                        callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.ListGraphs, handler.Uris), state);
-                    }
-                    else
-                    {
-                        callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.ListGraphs, args.Error), state);
-                    }
-                }, state);
-            }
-            else
-            {
-                callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.ListGraphs, new RdfStorageException("Underlying store does not supported listing graphs asynchronously or has failed to appropriately override this method")), state);
-            }
-        }
-
-        /// <summary>
-        /// Indicates whether the Store is ready to accept requests.
-        /// </summary>
-        public abstract bool IsReady
-        {
-            get;
-        }
-
-        /// <summary>
-        /// Gets whether the Store is read only.
-        /// </summary>
-        public abstract bool IsReadOnly
-        {
-            get;
-        }
-
-        /// <summary>
-        /// Gets the IO Behaviour of the Store.
-        /// </summary>
-        public abstract IOBehaviour IOBehaviour
-        {
-            get;
-        }
-
-        /// <summary>
-        /// Gets whether the Store supports Triple level updates via the <see cref="BaseAsyncHttpConnector.UpdateGraph(Uri,IEnumerable{Triple},IEnumerable{Triple},AsyncStorageCallback,Object)">UpdateGraph()</see> method.
-        /// </summary>
-        public abstract bool UpdateSupported
-        {
-            get;
-        }
-
-        /// <summary>
-        /// Gets whether the Store supports Graph deletion via the <see cref="BaseAsyncHttpConnector.DeleteGraph(Uri, AsyncStorageCallback, Object)">DeleteGraph()</see> method.
-        /// </summary>
-        public abstract bool DeleteSupported
-        {
-            get;
-        }
-
-        /// <summary>
-        /// Gets whether the Store supports listing graphs via the <see cref="BaseAsyncHttpConnector.ListGraphs(AsyncStorageCallback, Object)">ListGraphs()</see> method.
-        /// </summary>
-        public abstract bool ListGraphsSupported
-        {
-            get;
-        }
-
-        /// <summary>
-        /// Diposes of the Store.
-        /// </summary>
-        public abstract void Dispose();
-
-        /// <summary>
-        /// Helper method for doing async operations where a sequence of HTTP requests must be run.
-        /// </summary>
-        /// <param name="requests">HTTP requests.</param>
-        /// <param name="callback">Callback.</param>
-        /// <param name="state">State to pass to the callback.</param>
-        protected internal void MakeRequestSequence(IEnumerable<HttpWebRequest> requests, AsyncStorageCallback callback, Object state)
-        {
-            _d.BeginInvoke(requests, callback, state, MakeRequestSequenceCallback, callback);
-        }
-
-        private void MakeRequestSequenceCallback(IAsyncResult r)
-        {
-            AsyncStorageCallback callback = r.AsyncState as AsyncStorageCallback;
-            try
-            {
-                _d.EndInvoke(r);
-                // callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.Unknown), null);
-            }
-            catch (Exception ex)
-            {
-                if (callback != null)
-                {
-                    callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.Unknown, new RdfStorageException("Unexpected error while making a sequence of asynchronous requests to the Store, see inner exception for details", ex)), null);
-                }
-            }
-        }
-
-        private delegate void DoRequestSequenceDelgate(IEnumerable<HttpWebRequest> requests, AsyncStorageCallback callback, Object state);
-
-        private void DoRequestSequence(IEnumerable<HttpWebRequest> requests, AsyncStorageCallback callback, Object state)
-        {
-            ManualResetEvent signal = new ManualResetEvent(false);
-            foreach (HttpWebRequest request in requests)
-            {
-                Tools.HttpDebugRequest(request);
-
-                request.BeginGetResponse(r =>
-                {
-                    try
-                    {
-                        HttpWebResponse response = (HttpWebResponse)request.EndGetResponse(r);
-
-                        // This request worked OK, close the response and carry on
-                        response.Close();
-                        signal.Set();
-                    }
-                    catch (WebException webEx)
-                    {
-                        if (webEx.Response != null) Tools.HttpDebugResponse((HttpWebResponse)webEx.Response);
-                        callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.Unknown, new RdfStorageException("A HTTP Error occurred while making a sequence of asynchronous requests to the Store, see inner exception for details", webEx)), state);
-                        return;
-                    }
-                    catch (Exception ex)
-                    {
-                        callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.Unknown, new RdfStorageException("Unexpected error while making a sequence of asynchronous requests to the Store, see inner exception for details", ex)), state);
-                        return;
-                    }
-                }, state);
-
-                signal.WaitOne();
-                signal.Reset();
-            }
-
-            callback(this, new AsyncStorageCallbackArgs(AsyncStorageOperation.Unknown), state);
+            _hasCredentials = (!string.IsNullOrEmpty(username) && !string.IsNullOrEmpty(password));
         }
     }
 }
